Provide more useful j.l.Thread#getStackTrace stub (#1899)
The "empty trace" implementation of this commit is not as useful as
a proper, full implementation, but it is good enough for some
purposes, and is technically allowed by the spec.

ThreadTest.scala (JUnit) was ported from Scala.js. It was lightly
edited to be applicable to current Scala Native. It exercises the
change introduced in this commit to verify that it is effective.