Upgrade build to sbt 1.4.1 (#1956)
sbt 1.4.0 introduces mandatory checking for apparently unused
keys.  It uses an algorithm which may or may not be accurate.
sbt 1.4.0 complains about the key 'crossSbtVersions' in several
sub-projects.  sbt says that it is unused; lore suggests that it
is used in a subtle but critical way.  I did not have time to
verify the lore, so I turned linting off for that key.  The situation
is no worse than before. Yes, perhaps I missed an opportunity to
get rid of a truly unseen key, but I did not want to inflict
breakage on others. Not a way to win friends.